+++ cr-20181201/alibabacloud_cr20181201/models/_route_item.py
@@ -11,10 +11,16 @@ class RouteItem(DaraModel):
         instance_domain: str = None,
         storage_domain: str = None,
     ):
+        # Instance Endpoint Type
+        # 
         # This parameter is required.
         self.endpoint_type = endpoint_type
+        # Instance Domain
+        # 
         # This parameter is required.
         self.instance_domain = instance_domain
+        # OSS Domain
+        # 
         # This parameter is required.
         self.storage_domain = storage_domain